Add guards to unpatch(), fix failing patch tests
