--- tests/contrib/kafka/test_kafka_patch_generated.py
@@ -1,26 +1,0 @@
-# This test script was automatically generated by the contrib-patch-tests.py
-# script. If you want to make changes to it, you should make sure that you have
-# removed the ``_generated`` suffix from the file name, to prevent the content
-# from being overwritten by future re-generations.
-
-from ddtrace.contrib.kafka.patch import patch
-from tests.contrib.patch import PatchTestCase
-
-
-class TestKafkaPatch(PatchTestCase.Base):
-    __integration_name__ = "kafka"
-    __module_name__ = "confluent_kafka"
-    __patch_func__ = patch
-    __unpatch_func__ = None
-
-    # DEV: normally, we directly patch methods, but since confluent-kafka's methods are implemented in C and
-    # directly imported, our patched classes inherit from the original classes instead of ObjectProxy. This means
-    # regular patch test cases will not work.
-    def assert_module_patched(self, confluent_kafka):
-        pass
-
-    def assert_not_module_patched(self, confluent_kafka):
-        pass
-
-    def assert_not_module_double_patched(self, confluent_kafka):
-        pass
